+++ src/main/java/com/cryptodiff/marketMaps/MarketMaps.java
@@ -2,6 +2,7 @@ package com.cryptodiff.marketMaps;
 
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.stream.Collectors;
 
@@ -9,9 +10,9 @@ import java.util.stream.Collectors;
 public abstract class MarketMaps {
 
 
-    abstract HashMap<String, Double> getMarketPrices();
+    abstract HashMap<String, BigDecimal> getMarketPrices();
 
-    HashMap<String, Double> getOnlyUsdtPrices(HashMap<String, Double> marketPrices, String currency) {
+    HashMap<String, BigDecimal> getOnlyUsdtPrices(HashMap<String, BigDecimal> marketPrices, String currency) {
         return marketPrices.keySet().stream().filter(k -> k.contains(currency)).collect(Collectors.toMap(k -> k.replace(currency, "").toLowerCase(), v -> marketPrices.get(v), (v1, v2) -> v1, HashMap::new));
 
     }
